Add a docker image capable of producing a basic initramfs archive file.
This changeset introduces a Dockerfile that allows the User to create a
initramfs 'builder' image. The builder image will construct a barebones
initramfs archive that can be extracted via 'docker cp'. The intent is
to pass the initramfs archive as the argument to QEMU's -initrd option
when booting a custom kernel via the emulator. The 2021 blog post
"How To Build A Custom Linux Kernel For Qemu Using Docker" by Mitchel
Humpherys was the source of many of the commands used in this
Dockerfile. See
https://mgalgs.io/2021/03/23/how-to-build-a-custom-linux-kernel-for-qemu-using-docker.html