Simplify build.sh by moving the kernel config logic to kbuild.sh.
This changeset pushed the option of kernel configuration down into the
kbuild.sh script. The latter simplifies the logic in build.sh (i.e., now we
only have to run the kernel builder and initramfs Docker containers).
With this change, new ENV variables were added to the kernel builder
image:

  * KERNEL_SRC_DIR
  * KERNEL_OBJ_DIR
  * MODULE_SRC_DIR

The MODULE_SRC_DIR variable in particular was put in place in
anticipation of a future update which introduces custom module code into
the build process.