Make the kernel and module builds optional.
There's cases where you just want to build the kernel or just want to
build the modules. Prior to this changeset, the kbuild.sh script forced
the user to do both everytime. To add flexibility and speedup certain
builds, kbuild.sh prompts the user whether they want to build the kernel
and/or the modules.