Add module kernel objects to the initramfs.
This changeset ensures that the modules directory is included in the
Docker build context of the initramfs image. The Dockerfile of the initramfs
takes the modules directory and adds it to the filesystem. The end result is
that when you boot the kernel with the initramfs in QEMU, you can access the
modules directory which contains the *.ko. From there, you can
insmod/rmmod the module.