Add module params to the hello module.
This changeset introduces two parameters to the hello module: whom and
howmany. whom is a character string and howmany is an integer. The end
result is when the hello module is loaded and the params are supplied to
insmod, "Hello, <whom>" is printed howmany times. Try it out:

insmod hello.ko whom=Morbius howmany=3